Many changes since the last commit.  Some small refactoring w/ names of a few classes. The format in which the bot prints out the event entries has been revised.  Currently, create, edit, destroy, announce, and help commands are at least partially implemented (all usable). The schedule itself supports a title, a start time, an end time, and comments.
Next, the event entries must support dates (or at least days of weeks) as well as a 'time til starts' timer that updates periodically.  Current commands need to be fully implemented.  Logging needs to be implemented.  Recovery on disconnect or crashing needs to implemented.  And there is of course more than mentioned.