change: reenabled auto-sort; reduced thread pool used by bot shards; changed event announcement processing to use a fixed number of threads;
The theory behind *reducing* concurrency is that perhaps the connections opened to local database is becoming saturated, resulting in the bot locking up. This could explain some of the unusual behavior in which the bot appears to still be able to process event creation commands (and send an associated response) while the event is never saved to the db.